+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -2,7 +2,6 @@ import { Component, Renderer2 } from '@angular/core';
 import { CartService } from 'src/app/cart/services/order.service';
 import { ProductModel } from 'src/app/product/models/product.model';
 import { CartDTOModel } from '../../models/cart-dto.model';
-import { OrderByPipe } from 'src/app/shared/pipes/order-by.pipe';
 
 @Component({
   selector: 'app-cart-list',
@@ -12,8 +11,7 @@ import { OrderByPipe } from 'src/app/shared/pipes/order-by.pipe';
 export class CartListComponent {
 
   constructor(
-    private orderService: CartService,
-    private renderer: Renderer2
+    private orderService: CartService
   ) {
   }